requirements.txt
streamlit
pandas
pdfplumber
fpdf2